#7 namespace refactoring: phase 1.
