+++ src/Reader/Filter/FilterProcessorInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Reader\Filter\Processor;
+namespace Yiisoft\Data\Reader\Filter;
 
 
 interface FilterProcessorInterface
