+++ src/Reader/Iterable/Processor/GroupProcessor.php
@@ -1,10 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Reader\Filter\Processor\Iterable;
+namespace Yiisoft\Data\Reader\Iterable\Processor;
 
 
-use Yiisoft\Data\Reader\Filter\Processor\FilterProcessorInterface;
+use Yiisoft\Data\Reader\Filter\FilterProcessorInterface;
 
 abstract class GroupProcessor implements IterableProcessorInterface, FilterProcessorInterface
 {
