+++ src/Reader/Iterable/Processor/In.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Reader\Filter\Processor\Iterable;
+namespace Yiisoft\Data\Reader\Iterable\Processor;
 
-use Yiisoft\Data\Reader\Filter\Processor\FilterProcessorInterface;
+use Yiisoft\Data\Reader\Filter\FilterProcessorInterface;
 
 class In implements IterableProcessorInterface, FilterProcessorInterface
 {
